chore: install cmake 3.22 in dev container
Although the projects only requires cmake >= 3.14 , version 3.20+
provides support for CMakePresets.json which makes multi-config builds
and integration in IDEs much cooler. Even if the container used the latest
version of cmake, command line usage of the CmakeLists.txt can still work
fine with 3.14 although with the usual pain of passing tons of variables to
cmake and managing the cmake cache.