chore: configure valgrind with a specific ASAP_WITH_VALGRIND option
Using ${PROJECT_NAME}_ENABLE_PROFILING is too wide. To have fine grained
control over what profiling tools we are adding, we define a specific variable for
each one of them and we set the internal one accordingly. Only these
fine-grained controls are exposed to the user and the internal one is now
marked as advanced.