fix: remove axiom from the contrcat check apis
The implementation of axiom is still very fragile in different compilers.
Due to the lack of consistency between compilers when it comes to support
of `assume`, `builtin_unreachable` and the behavior when unreachable is
encountered, it is better to not offer this as part of the contract checking
apis for now.