perf: restrict exceptions to error reporting
BREAKING CHANGE:
Using exceptions to report completion or events not being consumed results
in significant overhead on the FSM due to exception handling and a loss of
performance. Changes were made to the API and implementation to make state
`OnLeave` and `OnEnter` return a Status and to let the action `Execute`
method also return a status.