refactor: set visibility to hidden by default
By default, we now set C++ visibility to hidden, and then we explicitly export
symbols that we want visible. Just be aware that for throwable user types, it
is important that the class does not have any members using non exported classes.
The best way to deal with this if the need to have complex members arises is
to use a pimpl.