refactor: trim the project to the bare minimum
The intention from the asap base project is to be a template starter
repo that can be cloned and used for new projects. The bare minimum is
the cmake build infrastructure, documentation, the common submodule and
the contract checking assertions submodule.

Contract checking library is now in a separate submodule to keep the
common module as minimal as possible. Projects that do not intend to use
contract checking will not have to depend on it.