fix: always use top level build dir for output
Even if the project is used as a sub-project, we want all output artifacts to
be placed in the top level build dir so that we can run tests, collect install
artifacts, etc... from a single location. THis would work if the project is used
as a top-level or as a sub-project.