fix: use cmake-format extension default behavior
By default the cmake-format extension will run the cmake-format command
in the workspace root and cmake-format will look for a usable
configuration file in the hierarchy of the file being formatted. Our
config file will be detected and used with no explicit setting here. Forcing
the setting here will produce an error with cmake-format on windows.