fix: generated `version.h` should follow project naming
It is preferred that this file is generated with always the same name
(`version.h`) but placed in a directory that follows the project meta
info. It is now placed in a directory under include named with the
project ID in lower case. See the `CMakeLists.txt` for how the project
ID is derived from its name.