feat: add the `version-info` tool to print the project info
This tool uses the generated `version.h` file in a small C++ program
to print the project's info, as defined in the project's master
`CMakeLists.txt`.

It also constitutes an example of how to use the `version.h` file and
a simple test to check that the `asap` infrastructure for defining and
building targets is working.