feat: enhance configure logs with project/module nesting hierarchy
Track the projects/modules nesting level with a hierarchy stack updated
when we enter/exit a project/module. Most of the management is done
automatically as helper functions get called to add modules or external
packages. Some of the boilerplate (minimal) is still manual:

* In the top-level `CMakeLists.txt`, the project needs to pushed at the
beginning and popped at the end.
* In each module `CMakeLists.txt`, the module needs to be pushed at
the start and popped at the end.

Use the `ASAP_LOG_PROJECT_HIERARCHY` to get a string that contains
the nesting hierarchy.