fix(build): #21 target option `WARNING` not propagated properly
By default, for every target we build, a compiler option will be added
to treat warnings as errors, unless the target is added with `WARNING` as
an option.

When that option is used, we now properly propagate it to set the
corresponding compiler option to **NOT** treat warnings as errors.

This is useful when 3rd party dependencies have include files that
generate warnings.